+++ EarthAsylumConsulting/Required/eacDoojigger_registration/eacDoojigger_registration.interface.php
@@ -9,8 +9,8 @@ namespace EarthAsylumConsulting\Interfaces;
  * @category	WordPress Plugin
  * @package		{eac}SoftwareRegistry
  * @author		Kevin Burkholder <KBurkholder@EarthAsylum.com>
- * @copyright	Copyright (c) 2025 EarthAsylum Consulting <www.EarthAsylum.com>
- * @version		25.0725.1
+ * @copyright	Copyright (c) 2026 EarthAsylum Consulting <www.EarthAsylum.com>
+ * @version		26.0729.1
  */
 
 interface eacDoojigger_registration
@@ -106,7 +106,7 @@ interface eacDoojigger_registration
 	 * @param string $endpoint one of 'create', 'activate', 'deactivate', 'verify'
 	 * @return string
 	 */
-	public function getApiEndPoint(string $endpoint = null);
+	public function getApiEndPoint(?string $endpoint = null);
 
 
 	/**